Fixes #43 - refactor app.go into separate files/packages
