+++ build_samples.sh
@@ -1,5 +1,5 @@
 #!/bin/sh
 
-./faas-cli -action build -yaml ./samples.yml
+./faas-cli -action build -yaml ./samples.yml # -squash=true
 
 docker images |head -n 4
