+++ builder/exec.go
@@ -1,0 +1,16 @@
+package builder
+
+import (
+	"os"
+	"os/exec"
+)
+
+// ExecCommand run a system command
+func ExecCommand(tempPath string, builder []string) {
+	targetCmd := exec.Command(builder[0], builder[1:]...)
+	targetCmd.Dir = tempPath
+	targetCmd.Stdout = os.Stdout
+	targetCmd.Stderr = os.Stderr
+	targetCmd.Start()
+	targetCmd.Wait()
+}
